+++ lifelines/fitters/log_normal_fitter.py
@@ -175,7 +175,7 @@ class LogNormalFitter(UnivariateFitter):
             warnings.simplefilter("ignore")
 
             results = minimize(
-                value_and_grad(_negative_log_likelihood), # pylint: disable=no-value-for-parameter
+                value_and_grad(_negative_log_likelihood),  # pylint: disable=no-value-for-parameter
                 initial_values,
                 jac=True,
                 method="CG",
@@ -184,7 +184,10 @@ class LogNormalFitter(UnivariateFitter):
             )
 
             if results.success:
-                hessian_ = hessian(_negative_log_likelihood)(results.x, log(T), E) # pylint: disable=no-value-for-parameter
+                # pylint: disable=no-value-for-parameter
+                hessian_ = hessian(_negative_log_likelihood)(
+                    results.x, log(T), E
+                )  
                 return results.x, -results.fun, hessian_ * T.shape[0]
             print(results)
             raise ConvergenceError("Did not converge. This is a lifelines problem, not yours;")