+++ lifelines/fitters/weibull_fitter.py
@@ -173,17 +173,17 @@ class WeibullFitter(UnivariateFitter):
             warnings.simplefilter("ignore")
 
             results = minimize(
-                value_and_grad(_negative_log_likelihood), # pylint: disable=no-value-for-parameter
+                value_and_grad(_negative_log_likelihood),  # pylint: disable=no-value-for-parameter
                 initial_values,
                 jac=True,
                 method="L-BFGS-B",
                 args=(T, E),
-                bounds=((0.000001, None), (0.000001, None)), # to stay well away from 0.
+                bounds=((0.000001, None), (0.000001, None)),  # to stay well away from 0.
                 options={"disp": show_progress},
             )
 
             if results.success:
-                hessian_ = hessian(_negative_log_likelihood)(results.x, T, E) # pylint: disable=no-value-for-parameter
+                hessian_ = hessian(_negative_log_likelihood)(results.x, T, E)  # pylint: disable=no-value-for-parameter
                 return results.x, -results.fun, hessian_ * T.shape[0]
             print(results)
             raise ConvergenceError("Did not converge. This is a lifelines problem, not yours;")